added some debug info to the run_bind script

+++ docker scripts/run_bind.bash
@@ -9,13 +9,57 @@ if [ "$#" -ne 1 ]; then
 fi
 DEBUG=$1
 
-pkill named
-pkill tcpdump
-cd /tmp
-if [ "$DEBUG" = "true" ]; then
-    echo "DEBUG MODE"
-    tcpdump -i any -w /tmp/$ALG-ns-root.pcap &
-    gdb --batch -ex "run" -ex "bt" -ex "quit" --args named -g -d 10
+# print information before running
+version_file=/OQS-bind/VERSION
+version=0
+if [[ -f $version_file ]]; then
+    version=$(<$version_file)
+fi
+echo "bind version: $version"
+
+namedconf_file="/usr/local/etc/named.conf"
+udp_mode=NONE
+if [[ -f $version_file ]]; then
+    udp_mode=$(awk '/udp_fragmentation/ {print $2}' /usr/local/etc/named.conf | tr -d ';')
+fi
+echo "udp fragmentation: $udp_mode"
+
+cipher=NONE
+while read -r file; do
+    FILE_ALG=$(sed -n '2p' "$file" | awk -F'[()]' '{print $2}') 
+    # check if key file exists
+    key_file="${file%.private}.key"
+    if [ ! -f "$key_file" ]; then
+        echo "Error: Cannot find '$key_file'"
+        exit 1
+    fi
+    # read keyfile
+    first_line=$(sed -n '1p' "$key_file")
+    if [[ "$first_line" == *"This is a key-signing key"* ]]; then
+        echo "KSK Algorithm: $FILE_ALG"
+    elif [[ "$first_line" == *"This is a zone-signing key"* ]]; then
+        echo "ZSK Algorithm: $FILE_ALG"
+    else
+        echo "Error: '$key_file' is neither a KSK or ZSK"
+        exit 1
+    fi
+done < <(find "$dir" -type f -name "/usr/local/etc/bind/zones/K*.private")
+echo "cipher: $cipher"
+read -p "do you want to run bind with these settings? (Y/N): " choice
+
+# Check the user's input
+if [[ "$choice" =~ ^[Yy]$ ]]; then
+    pkill named
+    pkill tcpdump
+    cd /tmp
+    if [ "$DEBUG" = "true" ]; then
+        echo "DEBUG MODE"
+        tcpdump -i any -w /tmp/$ALG-ns-root.pcap &
+        gdb --batch -ex "run" -ex "bt" -ex "quit" --args named -g -d 10
+    else
+        named -g -d 3
+    fi
 else
-    named -g -d 3
+    echo "aborting..."
+    exit 1
 fi
